add everything to the index.ts
